==> scripts/04_freqDist.ipynb <==
# Character Shifts in Harry Potter Fanfics

# Cooccurring words

### Last updated: 19.01.2022

### 1. Required Libraries

import pandas as pd
import io
import csv
import glob
import re
import nltk
import spacy
from nltk import FreqDist
from nltk import word_tokenize

path_cooccurrences = r'Z:\Fanfiction\HP_Character-Distribution\pamphlet_character_shifts\results\results\cooccurrences'

stopwords = nltk.corpus.stopwords.words('german')
nlp = spacy.load('de_core_news_sm')
nlp.max_length = 1000000000

path_txt = path_cooccurrences + '\\ffs_pairs_sentences'
list_files = glob.glob(path_txt + '/*.txt')

all_texts = []
for file in list_files:
    filename = re.sub(r'.+?sentences\\(.+).txt', r'\1', file)
    
    with open(file, 'r', encoding='utf-8') as txt_file:
        doc = txt_file.readlines()
        
    all_texts.append(doc)

texts_in_sents = []
i = 0

for pair in all_texts:
    filename = re.sub(r'.+?sentences\\(.+).txt', r'\1', list_files[i])
    toknum = 0
    toks = []
    for sentence in pair:
        tokenized = word_tokenize(sentence)
        toknum += len(tokenized)
        for tok in tokenized:
            if tok not in stopwords and tok.isalpha():
                toks.append(tok)
    
    fdist = FreqDist(word for word in toks)
    df_fdist = pd.DataFrame.from_dict(fdist, orient='index')
    df_fdist.columns = ['relative_frequency']
    df_fdist.index.name = 'term'
    df_fdist['relative_frequency'] = df_fdist['relative_frequency'].div(toknum)
    
    df_fdist_filename = path_cooccurrences + '\\cooccurring_words\\ffs\\' + filename + '_fdist_verbs.csv'
    df_fdist.to_csv(df_fdist_filename)
    i += 1

path_txt = path_cooccurrences + '\\originals_pairs_sentences'
list_files = glob.glob(path_txt + '/*.txt')

all_texts = []
for file in list_files:
    filename = re.sub(r'.+?sentences\\(.+).txt', r'\1', file)
    
    with open(file, 'r', encoding='utf-8') as txt_file:
        doc = txt_file.readlines()
        
    all_texts.append(doc)

texts_in_sents = []
i = 0

for pair in all_texts:
    filename = re.sub(r'.+?sentences\\(.+).txt', r'\1', list_files[i])
    toknum = 0
    toks = []
    for sentence in pair:
        tokenized = word_tokenize(sentence)
        toknum += len(tokenized)
        for tok in tokenized:
            if tok not in stopwords and tok.isalpha():
                toks.append(tok)
    
    fdist = FreqDist(word for word in toks)
    df_fdist = pd.DataFrame.from_dict(fdist, orient='index')
    df_fdist.columns = ['relative_frequency']
    df_fdist.index.name = 'term'
    df_fdist['relative_frequency'] = df_fdist['relative_frequency'].div(toknum)
    
    df_fdist_filename = path_cooccurrences + '\\cooccurring_words\\originals\\' + filename + '_fdist_verbs.csv'
    df_fdist.to_csv(df_fdist_filename)
    i += 1
