==> scripts/05_Word2Vec.ipynb <==
# Character Shifts in Harry Potter Fanfics

# Creation of Word Embedding Models

### Last updated: 19.01.2022

### 1. Required Libraries

import glob
import os
import nltk
import spacy
!python -m spacy download de_core_news_lg
nlp = spacy.load('de_core_news_lg',exclude=["ner"],disable=["tagger","parser"])
import string
import csv
from gensim.models import Word2Vec
import pickle

path_models = r'Z:\Fanfiction\HP_Character-Distribution\pamphlet_character_shifts\results\vector_models'
path_pickled = r'Z:\Fanfiction\HP_Character-Distribution\pamphlet_character_shifts\results\pickled'

### 2. Read in data

with open(path_pickled + '\\corpusHPoriginals_words.pkl', 'rb') as f:
    corpusHPoriginals = pickle.load(f)

with open(path_pickled + '\\corpusHPFFs_words.pkl', 'rb') as f:
    corpusHPFFs = pickle.load(f)

### 3. Preprocessing

already covered by "generaliseEntities":
- removal paragraph markers
- lowercase the corpus
- generalising entities

done here:
- merge texts in corpus into two large lists
- clean data (remove quotation marks)
- tokenizing sentences
- lemmatising
- removing punctuation (no "_"!!!)
- saving files

# merge texts in corpus into two large lists

corpusHPoriginals_joined = ' '.join(corpusHPoriginals)

corpusHPFFs_joined = ' '.join(corpusHPFFs)

# clean data (remove quotation marks)

corpusHPoriginals_clean1 = corpusHPoriginals_joined.replace('“', ' ')

corpusHPoriginals_clean2 = corpusHPoriginals_clean1.replace('„', ' ')

corpusHPFFs_clean1 = corpusHPFFs_joined.replace('“', ' ')

corpusHPFFs_clean2 = corpusHPFFs_clean1.replace('„', ' ')

# tokenizing sentences

corpusHPoriginals_sentences = nltk.sent_tokenize(corpusHPoriginals_clean2, language='german')

corpusHPFFs_sentences = nltk.sent_tokenize(corpusHPFFs_clean2, language='german')

# divide HPFFs corpus into more manageable chunks

len(corpusHPFFs_sentences)

corpusHPFFs_sentences1 = corpusHPFFs_sentences[:2776735]

corpusHPFFs_sentences2 = corpusHPFFs_sentences[2776735:5553470]

corpusHPFFs_sentences3 = corpusHPFFs_sentences[5553470:]

# lemmatising

# books

corpusHPoriginals_lemmatized = [0]*len(corpusHPoriginals_sentences)

for i in range(0, len(corpusHPoriginals_sentences)):
    words = nlp(corpusHPoriginals_sentences[i])
    interim = [0]*len(words)
    for j in range(0, len(interim)):
        interim[j] = words[j].lemma_
    corpusHPoriginals_lemmatized[i] = interim 

corpusHPoriginals_lemmatized[:10]

corpusHPoriginals_lemmatized[-10:]

corpusHPFFs_lemmatized1 = [0]*len(corpusHPFFs_sentences1)

for i in range(0, len(corpusHPFFs_sentences1)):
    words = nlp(corpusHPFFs_sentences1[i])
    interim = [0]*len(words)
    for j in range(0, len(interim)):
        interim[j] = words[j].lemma_
    corpusHPFFs_lemmatized1[i] = interim 

corpusHPFFs_lemmatized2 = [0]*len(corpusHPFFs_sentences2)

for i in range(0, len(corpusHPFFs_sentences2)):
    words = nlp(corpusHPFFs_sentences2[i])
    interim = [0]*len(words)
    for j in range(0, len(interim)):
        interim[j] = words[j].lemma_
    corpusHPFFs_lemmatized2[i] = interim 

corpusHPFFs_lemmatized3 = [0]*len(corpusHPFFs_sentences3)

for i in range(0, len(corpusHPFFs_sentences3)):
    words = nlp(corpusHPFFs_sentences3[i])
    interim = [0]*len(words)
    for j in range(0, len(interim)):
        interim[j] = words[j].lemma_
    corpusHPFFs_lemmatized3[i] = interim 

corpusHPFFs_lemmatized1[:10]

corpusHPFFs_lemmatized3[-10:]

# merge lemmatized HPFF corpus

corpusHPFFs_lemmatized = corpusHPFFs_lemmatized1 + corpusHPFFs_lemmatized2 + corpusHPFFs_lemmatized3

corpusHPFFs_lemmatized[:10]

corpusHPFFs_lemmatized[-10:]

# removing punctuation (no "_"!!!)

punctuation = """!"#$%&'()*+,-./:;<=>?@[\]^`{|}~«»"""

for sent in corpusHPoriginals_lemmatized:
    for word in sent:
        if word in punctuation:
            sent.remove(word)

for sent in corpusHPFFs_lemmatized:
    for word in sent:
        if word in punctuation:
            sent.remove(word)

# saving files

with open(path_pickled + '\\hp_originals_texts_lemmatized.pkl', 'wb') as f:
    pickle.dump(corpusHPoriginals_lemmatized, f)
    
#with open('hp_originals_texts_lemmatized.pkl', 'rb') as f:
#    corpusHPoriginals_lemmatized = pickle.load(f)

#with open('hp_ffs_texts_lemmatized.pkl', 'wb') as f:
#    pickle.dump(corpusHPFFs_lemmatized, f)
    
with open(path_pickled + '\\hp_ffs_texts_lemmatized.pkl', 'rb') as f:
    corpusHPFFs_lemmatized = pickle.load(f)

# saving files as csv

with open(path_results + '\\hp_originals_texts_lemmatized.csv', 'w', newline='') as file:
    writer = csv.writer(file)
    writer.writerows(corpusHPoriginals_lemmatized)

with open(path_results + '\\hp_ffs_texts_lemmatized.csv', 'w', newline='') as file:
    writer = csv.writer(file)
    writer.writerows(corpusHPFFs_lemmatized)

## 3. Creating Models

# training models

modelHPoriginalsA = Word2Vec(corpusHPoriginals_lemmatized, vector_size=300, window=5, min_count=50, sg=0, epochs=5)

modelHPoriginalsB = Word2Vec(corpusHPoriginals_lemmatized, vector_size=300, window=5, min_count=50, sg=0, epochs=10)

modelHPoriginalsC = Word2Vec(corpusHPoriginals_lemmatized, vector_size=300, window=5, min_count=50, sg=1, epochs=5)

modelHPoriginalsD = Word2Vec(corpusHPoriginals_lemmatized, vector_size=300, window=5, min_count=50, sg=1, epochs=10)

modelHPoriginalsE = Word2Vec(corpusHPoriginals_lemmatized, vector_size=300, window=5, min_count=50, sg=1, epochs=15)

modelHPoriginalsF = Word2Vec(corpusHPoriginals_lemmatized, vector_size=300, window=5, min_count=50, sg=1, epochs=20)

modelHPoriginalsG = Word2Vec(corpusHPoriginals_lemmatized, vector_size=300, window=5, min_count=50, sg=1, epochs=25)

modelHPoriginalsH = Word2Vec(corpusHPoriginals_lemmatized, vector_size=200, window=5, min_count=50, sg=1, epochs=5)

modelHPoriginalsI = Word2Vec(corpusHPoriginals_lemmatized, vector_size=200, window=5, min_count=50, sg=1, epochs=10)

modelHPoriginalsJ = Word2Vec(corpusHPoriginals_lemmatized, vector_size=200, window=5, min_count=50, sg=1, epochs=15)

modelHPoriginalsK = Word2Vec(corpusHPoriginals_lemmatized, vector_size=200, window=5, min_count=50, sg=1, epochs=20)

modelHPoriginalsL = Word2Vec(corpusHPoriginals_lemmatized, vector_size=200, window=5, min_count=50, sg=1, epochs=25)

modelHPoriginalsM = Word2Vec(corpusHPoriginals_lemmatized, vector_size=100, window=5, min_count=50, sg=1, epochs=5)

modelHPoriginalsN = Word2Vec(corpusHPoriginals_lemmatized, vector_size=100, window=5, min_count=50, sg=1, epochs=10)

modelHPoriginalsO = Word2Vec(corpusHPoriginals_lemmatized, vector_size=100, window=5, min_count=50, sg=1, epochs=15)

modelHPoriginalsP = Word2Vec(corpusHPoriginals_lemmatized, vector_size=100, window=5, min_count=50, sg=1, epochs=20)

modelHPoriginalsQ = Word2Vec(corpusHPoriginals_lemmatized, vector_size=100, window=5, min_count=50, sg=1, epochs=25)


# saving models

modelHPoriginalsA.wv.save(path_models + '\\modelHPoriginalsA_vectors.kv')

modelHPoriginalsB.wv.save(path_models + '\\modelHPoriginalsB_vectors.kv')

modelHPoriginalsC.wv.save(path_models + '\\modelHPoriginalsC_vectors.kv')

modelHPoriginalsD.wv.save(path_models + '\\modelHPoriginalsD_vectors.kv')

modelHPoriginalsE.wv.save(path_models + '\\modelHPoriginalsE_vectors.kv')

modelHPoriginalsF.wv.save(path_models + '\\modelHPoriginalsF_vectors.kv')

modelHPoriginalsG.wv.save(path_models + '\\modelHPoriginalsG_vectors.kv')

modelHPoriginalsH.wv.save(path_models + '\\modelHPoriginalsH_vectors.kv')

modelHPoriginalsI.wv.save(path_models + '\\modelHPoriginalsI_vectors.kv')

modelHPoriginalsJ.wv.save(path_models + '\\modelHPoriginalsJ_vectors.kv')

modelHPoriginalsK.wv.save(path_models + '\\modelHPoriginalsK_vectors.kv')

modelHPoriginalsL.wv.save(path_models + '\\modelHPoriginalsL_vectors.kv')

modelHPoriginalsM.wv.save(path_models + '\\modelHPoriginalsM_vectors.kv')

modelHPoriginalsN.wv.save(path_models + '\\modelHPoriginalsN_vectors.kv')

modelHPoriginalsO.wv.save(path_models + '\\modelHPoriginalsO_vectors.kv')

modelHPoriginalsP.wv.save(path_models + '\\modelHPoriginalsP_vectors.kv')

modelHPoriginalsQ.wv.save(path_models + '\\modelHPoriginalsQ_vectors.kv')


# training models

modelHPFFsA = Word2Vec(corpusHPFFs_lemmatized, vector_size=300, window=5, min_count=50, sg=0, epochs=5)

modelHPFFsB = Word2Vec(corpusHPFFs_lemmatized, vector_size=300, window=5, min_count=50, sg=0, epochs=10)

modelHPFFsC = Word2Vec(corpusHPFFs_lemmatized, vector_size=300, window=5, min_count=50, sg=1, epochs=5)

modelHPFFsD = Word2Vec(corpusHPFFs_lemmatized, vector_size=300, window=5, min_count=50, sg=1, epochs=10)

modelHPFFsE = Word2Vec(corpusHPFFs_lemmatized, vector_size=300, window=5, min_count=50, sg=1, epochs=15)

modelHPFFsF = Word2Vec(corpusHPFFs_lemmatized, vector_size=300, window=5, min_count=50, sg=1, epochs=20)

modelHPFFsG = Word2Vec(corpusHPFFs_lemmatized, vector_size=300, window=5, min_count=50, sg=1, epochs=25)

modelHPFFsH = Word2Vec(corpusHPFFs_lemmatized, vector_size=200, window=5, min_count=50, sg=1, epochs=5)

modelHPFFsI = Word2Vec(corpusHPFFs_lemmatized, vector_size=200, window=5, min_count=50, sg=1, epochs=10)

modelHPFFsJ = Word2Vec(corpusHPFFs_lemmatized, vector_size=200, window=5, min_count=50, sg=1, epochs=15)

modelHPFFsK = Word2Vec(corpusHPFFs_lemmatized, vector_size=200, window=5, min_count=50, sg=1, epochs=20)

modelHPFFsL = Word2Vec(corpusHPFFs_lemmatized, vector_size=200, window=5, min_count=50, sg=1, epochs=25)

modelHPFFsM = Word2Vec(corpusHPFFs_lemmatized, vector_size=100, window=5, min_count=50, sg=1, epochs=5)

modelHPFFsN = Word2Vec(corpusHPFFs_lemmatized, vector_size=100, window=5, min_count=50, sg=1, epochs=10)

modelHPFFsO = Word2Vec(corpusHPFFs_lemmatized, vector_size=100, window=5, min_count=50, sg=1, epochs=15)

modelHPFFsP = Word2Vec(corpusHPFFs_lemmatized, vector_size=100, window=5, min_count=50, sg=1, epochs=20)

modelHPFFsQ = Word2Vec(corpusHPFFs_lemmatized, vector_size=100, window=5, min_count=50, sg=1, epochs=25)

# saving models

modelHPFFsA.wv.save(path_models + '\\modelHPFFsA_vectors.kv')

modelHPFFsB.wv.save(path_models + '\\modelHPFFsB_vectors.kv')

modelHPFFsC.wv.save(path_models + '\\modelHPFFsC_vectors.kv')

modelHPFFsD.wv.save(path_models + '\\modelHPFFsD_vectors.kv')

modelHPFFsE.wv.save(path_models + '\\modelHPFFsE_vectors.kv')

modelHPFFsF.wv.save(path_models + '\\modelHPFFsF_vectors.kv')

modelHPFFsG.wv.save(path_models + '\\modelHPFFsG_vectors.kv')

modelHPFFsH.wv.save(path_models + '\\modelHPFFsH_vectors.kv')

modelHPFFsI.wv.save(path_models + '\\modelHPFFsI_vectors.kv')

modelHPFFsJ.wv.save(path_models + '\\modelHPFFsJ_vectors.kv')

modelHPFFsK.wv.save(path_models + '\\modelHPFFsK_vectors.kv')

modelHPFFsL.wv.save(path_models + '\\modelHPFFsL_vectors.kv')

modelHPFFsM.wv.save(path_models + '\\modelHPFFsM_vectors.kv')

modelHPFFsN.wv.save(path_models + '\\modelHPFFsN_vectors.kv')

modelHPFFsO.wv.save(path_models + '\\modelHPFFsO_vectors.kv')

modelHPFFsP.wv.save(path_models + '\\modelHPFFsP_vectors.kv')

modelHPFFsQ.wv.save(path_models + '\\modelHPFFsQ_vectors.kv')

### 4. Analyzing Corpus

# amount of tokens

i = sum([len(sent) for sent in corpusHPoriginals_lemmatized])

print(i)

# amount of types (in models)

len(modelHPoriginalsA)

# amount of tokens

i = sum([len(sent) for sent in corpusHPFFs_lemmatized])

print(i)

# amount of types (in models)

len(modelHPFFsA)

# create word lists

HPoriginals_words = [word for sent in corpusHPoriginals_lemmatized for word in sent]

HPFFs_words = [word for sent in corpusHPFFs_lemmatized for word in sent]

# saving files as csv

with open(path_results + '\\hp_originals_texts_lemmatized_words.csv', 'w', newline='') as file:
    writer = csv.writer(file)
    writer.writerows([HPoriginals_words[index]] for index in range(0, len(HPoriginals_words)))

with open(path_results + '\\hp_ffs_texts_lemmatized_words.csv', 'w', newline='') as file:
    writer = csv.writer(file)
    writer.writerows([HPFFs_words[index]] for index in range(0, len(HPFFs_words)))

HPoriginals_words.count('HARRY_POTTER')

HPFFs_words.count('HARRY_POTTER')

### 5. Analyzing Models

modelHPoriginalsA.wv.most_similar('HARRY_POTTER')

modelHPFFsA.wv.most_similar('HARRY_POTTER')
