==> scripts/08_countFrequenciesAll.ipynb <==
# Character Shifts in Harry Potter Fanfics

# Frequencies for all Emotion Words

### Last updated: 19.01.2022

1. import BAWL-R
2. extract WORD, WORD_LOWER, WORD_CLASS, EMO_MEAN, EMO_STD, AROUSAL_MEAN, AROUSAL_STD as lists
3. create list ORIGINALS_FREQ and FFS_FREQ based on the list WORD_LOWER
4. create new xlsx-file with WORD, WORD_LOWER, WORD_CLASS, EMO_MEAN, EMO_STD, AROUSAL_MEAN, AROUSAL_STD + ORIGINALS_FREQ, FFS_FREQ

path_data = r'Z:\Fanfiction\HP_Character-Distribution\pamphlet_character_shifts\data'
path_results = r'Z:\Fanfiction\HP_Character-Distribution\pamphlet_character_shifts\results'

# 1. import BAWL-R

import pandas as pd

df = pd.read_excel(path_data + '\\bawl\\BAWL-R.xlsx')

# 2. extract WORD, WORD_LOWER, WORD_CLASS, EMO_MEAN, EMO_STD, AROUSAL_MEAN, AROUSAL_STD as lists

WORD = df['WORD'].tolist()

len(WORD)

WORD_LOWER = df['WORD_LOWER'].tolist()

WORD_CLASS = df['WORD_CLASS'].tolist()

EMO_MEAN = df['EMO_MEAN'].tolist()

EMO_STD = df['EMO_STD'].tolist()

AROUSAL_MEAN = df['AROUSAL_MEAN'].tolist()

AROUSAL_STD = df['AROUSAL_STD'].tolist()


# 3. create list ORIGINALS_FREQ and FFS_FREQ based on the list WORD_LOWER

import csv

with open(path_results + '\\hp_originals_texts_lemmatized_words.csv', newline='') as f:
    reader = csv.reader(f)
    hp_originals_words = list(reader)

with open(path_results + '\\hp_ffs_texts_lemmatized_words.csv', newline='') as f:
    reader = csv.reader(f)
    hp_ffs_words = list(reader)

# flatten lists

hp_originals_words = [word for sublist in hp_originals_words for word in sublist]

hp_ffs_words = [word for sublist in hp_ffs_words for word in sublist]

# create frequency lists

ORIGINALS_FREQ = []

for word in WORD_LOWER:
    ORIGINALS_FREQ.append(hp_originals_words.count(word))

FFS_FREQ = []

for word in WORD_LOWER:
    FFS_FREQ.append(hp_ffs_words.count(word))

# 4. create new xlsx-file with WORD, WORD_LOWER, WORD_CLASS, EMO_MEAN, EMO_STD, AROUSAL_MEAN, AROUSAL_STD + ORIGINALS_FREQ, FFS_FREQ

BAWLR_with_freqs = pd.DataFrame(
    {'WORD': WORD,
     'WORD_LOWER': WORD_LOWER,
     'WORD_CLASS ': WORD_CLASS ,
     'EMO_MEAN': EMO_MEAN,
     'EMO_STD': EMO_STD,
     'AROUSAL_MEAN': AROUSAL_MEAN,
     'AROUSAL_STD': AROUSAL_STD,
     'ORIGINALS_FREQ': ORIGINALS_FREQ,
     'FFS_FREQ': FFS_FREQ
    }
)

BAWLR_with_freqs.to_excel(path_data + '\\bawl\\BAWLR_with_freqs.xlsx')

BAWLR_with_freqs.to_csv(path_data + '\\bawl\\BAWLR_with_freqs.csv')
