==> scripts/09_BAWLcluster.ipynb <==
# Character Shifts in Harry Potter Fanfics

# BAWL Cluster

### Last updated: 19.01.2022

## 0. Preliminaries

### 0.1 Import/Load

Import Modules:

import numpy as np
import pickle
import pandas as pd
from itertools import cycle
from gensim.models import KeyedVectors
from sklearn.preprocessing import normalize
from sklearn.decomposition import PCA
from sklearn.cluster import AffinityPropagation as AF

Load KeyesVectors Harry Potter books/fanfiction:

path_data = r'Z:\Fanfiction\HP_Character-Distribution\pamphlet_character_shifts\data'
path_models = r'Z:\Fanfiction\HP_Character-Distribution\pamphlet_character_shifts\results\vector_models'
path_pickled = r'Z:\Fanfiction\HP_Character-Distribution\pamphlet_character_shifts\results\pickled'

model_books = KeyedVectors.load(path_models + '\\modelHPoriginalsD_vectors.kv')

model_ff = KeyedVectors.load(path_models + '\\modelHPFFsD_vectors.kv')

# 1. BAWL

Load BAWL with word frequency of at least 50

bawl = pd.read_csv(path_data + '\\bawl\\BAWLR_with_freqs.csv')
bawl_words = list(bawl[(bawl["ORIGINALS_FREQ"] >= 50) & (bawl["FFS_FREQ"] >= 50)]["WORD_LOWER"])

## 1.1 Originals

### 1.1.1 Valence

#### 1.1.1a High Valence

Define relevant valence words by threshold value

emo_mean = 2
emo_std = 1

high_valence_words = list(bawl[(bawl["ORIGINALS_FREQ"] >= 50) & (bawl["EMO_MEAN"] >= emo_mean) & (bawl["EMO_STD"] <= emo_std)]["WORD_LOWER"])

Define relevant valence words by ranking

top_n = 25
emo_std = 1

high_valence_words = list(bawl[(bawl["ORIGINALS_FREQ"] >= 50) & (bawl["EMO_STD"] <= emo_std)].sort_values("EMO_MEAN")["WORD_LOWER"])[-top_n:]

Get vectors for valence words

valence_words_index = [model_books.key_to_index[word] for word in high_valence_words]
valence_word_vectors = [model_books.get_vector(i) for i in valence_words_index]

Normalize vectors on unit circle

valence_word_cosine = np.array(normalize(valence_word_vectors,norm='l2'))

Cluster with Affinity Propagation on PCA

valence_cos_cluster = AF(affinity='euclidean')
valence_cos_cluster.fit(valence_word_cosine)

Tranform cluster to dictionary

valence_sem_groups = {}
for i in range(len(valence_cos_cluster.labels_)):
    label = valence_cos_cluster.labels_[i]
    if label not in valence_sem_groups.keys():
        valence_sem_groups[label] = []

    word = high_valence_words[i]
    valence_sem_groups[label].append(word)

print(valence_sem_groups)

Print relevant information of clusters

suggestion_values = []

letters = cycle('abcdefghijklmnopqrstuvxyz')
for k, let in zip(range(len(valence_cos_cluster.cluster_centers_)), letters):
    class_members = valence_cos_cluster.labels_ == k

    dist_list = []
    print(let + ": ITEMS " + str(list(np.array(high_valence_words)[class_members])))
    for member1 in np.array(high_valence_words)[class_members]:
        for member2 in np.array(high_valence_words)[class_members]:
            if member1 != member2:
                dist_list.append(model_books.similarity(member1,member2))
    if (dist_list != []):
        print(let + ": AVG SIMILARITY " + str(sum(dist_list) / len(dist_list)))
        
    valence_means = []
    valence_stds = []
    arousal_means = []
    arousal_stds = []
    for word in valence_sem_groups[k]:
        word_row = bawl[bawl["WORD_LOWER"] == word]
        valence_means.append(float(word_row["EMO_MEAN"]))
        valence_stds.append(float(word_row["EMO_STD"]))
        arousal_means.append(float(word_row["AROUSAL_MEAN"]))
        arousal_stds.append(float(word_row["AROUSAL_STD"]))

    print(let + ": MEAN VALENCE " + str(sum(valence_means) / len(valence_means)))
    print(let + ": MEAN VALENCE STD " + str(sum(valence_stds) / len(valence_stds)))
    print(let + ": MEAN AROUSAL " + str(sum(arousal_means) / len(arousal_means)))
    print(let + ": MEAN AROUSAL STD " + str(sum(arousal_stds) / len(arousal_stds)))
    
    try:
        suggestion_value = (sum(dist_list) / len(dist_list) + abs(sum(valence_means) / len(valence_means)) - (sum(valence_stds) / len(valence_stds)))*(1.05**(len(np.array(high_valence_words)[class_members])))
    except:
        suggestion_value = 0
    suggestion_values.append((suggestion_value,np.array(high_valence_words)[class_members]))    
    print(let + ": SUGGESTION VALUE " + str(suggestion_values[-1][0]))                        

Ranking for suggestion values

suggestion_values = sorted(suggestion_values, key=lambda tup: tup[0], reverse=True)
for rank in range(len(suggestion_values)):
    print(str(rank+1) + " " + str(list(suggestion_values[rank][1])))

high_val_originals = list(suggestion_values[0][1])
with open(path_pickled + "\\high_val_originals.pkl", "wb") as f:
    pickle.dump(high_val_originals, f)

#### 1.1.1b Low Valence

Define relevant valence words by threshold value

emo_mean = -2
emo_std = 1

low_valence_words = list(bawl[(bawl["ORIGINALS_FREQ"] >= 50) & (bawl["EMO_MEAN"] <= emo_mean) & (bawl["EMO_STD"] <= emo_std)]["WORD_LOWER"])

Define relevant valence words by ranking

bot_n = 25
emo_std = 1

low_valence_words = list(bawl[(bawl["ORIGINALS_FREQ"] >= 50) & (bawl["EMO_STD"] <= emo_std)].sort_values("EMO_MEAN")["WORD_LOWER"])[:bot_n]

Get vectors for valence words

valence_words_index = [model_books.key_to_index[word] for word in low_valence_words]
valence_word_vectors = [model_books.get_vector(i) for i in valence_words_index]

Normalize vectors on unit circle

valence_word_cosine = np.array(normalize(valence_word_vectors,norm='l2'))

Cluster with Affinity Propagation

valence_cos_cluster = AF(affinity='euclidean')
valence_cos_cluster.fit(valence_word_cosine)

Tranform cluster to dictionary

valence_sem_groups = {}
for i in range(len(valence_cos_cluster.labels_)):
    label = valence_cos_cluster.labels_[i]
    if label not in valence_sem_groups.keys():
        valence_sem_groups[label] = []

    word = low_valence_words[i]
    valence_sem_groups[label].append(word)
    
print(valence_sem_groups)

Print relevant information of clusters

suggestion_values = []

letters = cycle('abcdefghijklmnopqrstuvxyz')
for k, let in zip(range(len(valence_cos_cluster.cluster_centers_)), letters):
    class_members = valence_cos_cluster.labels_ == k

    dist_list = []
    print(let + ": ITEMS " + str(list(np.array(low_valence_words)[class_members])))
    for member1 in np.array(low_valence_words)[class_members]:
        for member2 in np.array(low_valence_words)[class_members]:
            if member1 != member2:
                dist_list.append(model_books.similarity(member1,member2))
    if (dist_list != []):
        print(let + ": AVG SIMILARITY " + str(sum(dist_list) / len(dist_list)))
        
    valence_means = []
    valence_stds = []
    arousal_means = []
    arousal_stds = []
    for word in valence_sem_groups[k]:
        word_row = bawl[bawl["WORD_LOWER"] == word]
        valence_means.append(float(word_row["EMO_MEAN"]))
        valence_stds.append(float(word_row["EMO_STD"]))
        arousal_means.append(float(word_row["AROUSAL_MEAN"]))
        arousal_stds.append(float(word_row["AROUSAL_STD"]))

    print(let + ": MEAN VALENCE " + str(sum(valence_means) / len(valence_means)))
    print(let + ": MEAN VALENCE STD " + str(sum(valence_stds) / len(valence_stds)))
    print(let + ": MEAN AROUSAL " + str(sum(arousal_means) / len(arousal_means)))
    print(let + ": MEAN AROUSAL STD " + str(sum(arousal_stds) / len(arousal_stds)))
    
    try:
        suggestion_value = (sum(dist_list) / len(dist_list) + abs(sum(valence_means) / len(valence_means)) - (sum(valence_stds) / len(valence_stds)))*(1.05**(len(np.array(high_valence_words)[class_members])))
    except:
        suggestion_value = 0
    suggestion_values.append((suggestion_value,np.array(low_valence_words)[class_members]))    
    print(let + ": SUGGESTION VALUE " + str(suggestion_values[-1][0]))

Ranking for suggestion values

suggestion_values = sorted(suggestion_values, key=lambda tup: tup[0], reverse=True)
for rank in range(len(suggestion_values)):
    print(str(rank+1) + " " + str(list(suggestion_values[rank][1])))

low_val_originals = list(suggestion_values[0][1])
with open(path_pickled + "\\low_val_originals.pkl", "wb") as f:
    pickle.dump(low_val_originals, f)

### 1.1.2 Arousal

#### 1.1.2a High Arousal

Define relevant arousal words by threshold value

arousal_mean = 2
arousal_std = 1

high_arousal_words = list(bawl[(bawl["ORIGINALS_FREQ"] >= 50) & (bawl["AROUSAL_MEAN"] >= arousal_mean) & (bawl["AROUSAL_STD"] <= arousal_std)]["WORD_LOWER"])

Define relevant arousal words by ranking

top_n = 25
arousal_std = 1

high_arousal_words = list(bawl[(bawl["ORIGINALS_FREQ"] >= 50) & (bawl["AROUSAL_STD"] <= arousal_std)].sort_values("AROUSAL_MEAN")["WORD_LOWER"])[-top_n:]

Get vectors for arousal words

arousal_words_index = [model_books.key_to_index[word] for word in high_arousal_words]
arousal_word_vectors = [model_books.get_vector(i) for i in arousal_words_index]

Normalize vectors on unit circle

arousal_word_cosine = np.array(normalize(arousal_word_vectors,norm='l2'))

Cluster with Affinity Propagation

arousal_cos_cluster = AF(affinity='euclidean')
arousal_cos_cluster.fit(arousal_word_cosine)

Tranform cluster to dictionary

arousal_sem_groups = {}
for i in range(len(arousal_cos_cluster.labels_)):
    label = arousal_cos_cluster.labels_[i]
    if label not in arousal_sem_groups.keys():
        arousal_sem_groups[label] = []

    word = high_arousal_words[i]
    arousal_sem_groups[label].append(word)
    
print(arousal_sem_groups)

Print relevant information of clusters

suggestion_values = []

letters = cycle('abcdefghijklmnopqrstuvxyz')
for k, let in zip(range(len(arousal_cos_cluster.cluster_centers_)), letters):
    class_members = arousal_cos_cluster.labels_ == k

    dist_list = []
    print(let + ": ITEMS " + str(list(np.array(high_arousal_words)[class_members])))
    for member1 in np.array(high_arousal_words)[class_members]:
        for member2 in np.array(high_arousal_words)[class_members]:
            if member1 != member2:
                dist_list.append(model_books.similarity(member1,member2))
    if (dist_list != []):
        print(let + ": AVG SIMILARITY " + str(sum(dist_list) / len(dist_list)))
        
    valence_means = []
    valence_stds = []
    arousal_means = []
    arousal_stds = []
    for word in arousal_sem_groups[k]:
        word_row = bawl[bawl["WORD_LOWER"] == word]
        valence_means.append(float(word_row["EMO_MEAN"]))
        valence_stds.append(float(word_row["EMO_STD"]))
        arousal_means.append(float(word_row["AROUSAL_MEAN"]))
        arousal_stds.append(float(word_row["AROUSAL_STD"]))

    print(let + ": MEAN VALENCE " + str(sum(valence_means) / len(valence_means)))
    print(let + ": MEAN VALENCE STD " + str(sum(valence_stds) / len(valence_stds)))
    print(let + ": MEAN AROUSAL " + str(sum(arousal_means) / len(arousal_means)))
    print(let + ": MEAN AROUSAL STD " + str(sum(arousal_stds) / len(arousal_stds)))
    
    try:
        suggestion_value = (sum(dist_list) / len(dist_list) + sum(arousal_means) / len(arousal_means) - (sum(arousal_stds) / len(arousal_stds)))*(1.05**(len(np.array(high_arousal_words)[class_members])))
    except:
        suggestion_value = 0
    suggestion_values.append((suggestion_value,np.array(high_arousal_words)[class_members]))    
    print(let + ": SUGGESTION VALUE " + str(suggestion_values[-1][0]))

Ranking for suggestion values

suggestion_values = sorted(suggestion_values, key=lambda tup: tup[0], reverse=True)
for rank in range(len(suggestion_values)):
    print(str(rank+1) + " " + str(list(suggestion_values[rank][1])))

high_arousal_originals = list(suggestion_values[0][1])
with open(path_pickled + "\\high_arousal_originals.pkl", "wb") as f:
    pickle.dump(high_arousal_originals, f)

#### 1.1.2b Low Arousal

Define relevant arousal words by threshold value

arousal_mean = 2
arousal_std = 1

low_arousal_words = list(bawl[(bawl["ORIGINALS_FREQ"] >= 50) & (bawl["AROUSAL_MEAN"] <= arousal_mean) & (bawl["AROUSAL_STD"] <= arousal_std)]["WORD_LOWER"])

Define relevant arousal words by ranking

bot_n = 25
arousal_std = 1

low_arousal_words = list(bawl[(bawl["ORIGINALS_FREQ"] >= 50) & (bawl["AROUSAL_STD"] <= arousal_std)].sort_values("AROUSAL_MEAN")["WORD_LOWER"])[:bot_n]

Get vectors for arousal words

arousal_words_index = [model_books.key_to_index[word] for word in low_arousal_words]
arousal_word_vectors = [model_books.get_vector(i) for i in arousal_words_index]

Normalize vectors on unit circle

arousal_word_cosine = np.array(normalize(arousal_word_vectors,norm='l2'))

Cluster with Affinity Propagation

arousal_cos_cluster = AF(affinity='euclidean')
arousal_cos_cluster.fit(arousal_word_cosine)

Tranform cluster to dictionary

arousal_sem_groups = {}
for i in range(len(arousal_cos_cluster.labels_)):
    label = arousal_cos_cluster.labels_[i]
    if label not in arousal_sem_groups.keys():
        arousal_sem_groups[label] = []

    word = low_arousal_words[i]
    arousal_sem_groups[label].append(word)
    
print(arousal_sem_groups)

Print relevant information of clusters

suggestion_values = []

letters = cycle('abcdefghijklmnopqrstuvxyz')
for k, let in zip(range(len(arousal_cos_cluster.cluster_centers_)), letters):
    class_members = arousal_cos_cluster.labels_ == k

    dist_list = []
    print(let + ": ITEMS " + str(list(np.array(low_arousal_words)[class_members])))
    for member1 in np.array(low_arousal_words)[class_members]:
        for member2 in np.array(low_arousal_words)[class_members]:
            if member1 != member2:
                dist_list.append(model_books.similarity(member1,member2))
    if (dist_list != []):
        print(let + ": AVG SIMILARITY " + str(sum(dist_list) / len(dist_list)))
        
    valence_means = []
    valence_stds = []
    arousal_means = []
    arousal_stds = []
    for word in arousal_sem_groups[k]:
        word_row = bawl[bawl["WORD_LOWER"] == word]
        valence_means.append(float(word_row["EMO_MEAN"]))
        valence_stds.append(float(word_row["EMO_STD"]))
        arousal_means.append(float(word_row["AROUSAL_MEAN"]))
        arousal_stds.append(float(word_row["AROUSAL_STD"]))

    print(let + ": MEAN VALENCE " + str(sum(valence_means) / len(valence_means)))
    print(let + ": MEAN VALENCE STD " + str(sum(valence_stds) / len(valence_stds)))
    print(let + ": MEAN AROUSAL " + str(sum(arousal_means) / len(arousal_means)))
    print(let + ": MEAN AROUSAL STD " + str(sum(arousal_stds) / len(arousal_stds)))
    
    try:
        suggestion_value = (sum(dist_list) / len(dist_list) + (4 - sum(arousal_means) / len(arousal_means)) - (sum(arousal_stds) / len(arousal_stds)))*(1.05**(len(np.array(low_arousal_words)[class_members])))
    except:
        suggestion_value = 0
    suggestion_values.append((suggestion_value,np.array(low_arousal_words)[class_members]))    
    print(let + ": SUGGESTION VALUE " + str(suggestion_values[-1][0]))

Ranking for suggestion values

suggestion_values = sorted(suggestion_values, key=lambda tup: tup[0], reverse=True)
for rank in range(len(suggestion_values)):
    print(str(rank+1) + " " + str(list(suggestion_values[rank][1])))

low_arousal_originals = list(suggestion_values[0][1])
with open(path_pickled + "\\low_arousal_originals.pkl", "wb") as f:
    pickle.dump(low_arousal_originals, f)

## 1.2 Fanfiction

### 1.2.1 Valence

#### 1.2.1a High Valence

Define relevant valence words by threshold value

emo_mean = 2
emo_std = 1

high_valence_words = list(bawl[(bawl["FFS_FREQ"] >= 50) & (bawl["EMO_MEAN"] >= emo_mean) & (bawl["EMO_STD"] <= emo_std)]["WORD_LOWER"])

Define relevant valence words by ranking

top_n = 25
emo_std = 1

high_valence_words = list(bawl[(bawl["FFS_FREQ"] >= 50) & (bawl["EMO_STD"] <= emo_std)].sort_values("EMO_MEAN")["WORD_LOWER"])[-top_n:]

Get vectors for valence words

valence_words_index = [model_ff.key_to_index[word] for word in high_valence_words]
valence_word_vectors = [model_ff.get_vector(i) for i in valence_words_index]

Normalize vectors on unit circle

valence_word_cosine = np.array(normalize(valence_word_vectors,norm='l2'))

Cluster with Affinity Propagation on PCA

valence_cos_cluster = AF(affinity='euclidean')
valence_cos_cluster.fit(valence_word_cosine)

Tranform cluster to dictionary

valence_sem_groups = {}
for i in range(len(valence_cos_cluster.labels_)):
    label = valence_cos_cluster.labels_[i]
    if label not in valence_sem_groups.keys():
        valence_sem_groups[label] = []

    word = high_valence_words[i]
    valence_sem_groups[label].append(word)

print(valence_sem_groups)

Print relevant information of clusters

suggestion_values = []

letters = cycle('abcdefghijklmnopqrstuvxyz')
for k, let in zip(range(len(valence_cos_cluster.cluster_centers_)), letters):
    class_members = valence_cos_cluster.labels_ == k

    dist_list = []
    print(let + ": ITEMS " + str(list(np.array(high_valence_words)[class_members])))
    for member1 in np.array(high_valence_words)[class_members]:
        for member2 in np.array(high_valence_words)[class_members]:
            if member1 != member2:
                dist_list.append(model_ff.similarity(member1,member2))
    if (dist_list != []):
        print(let + ": AVG SIMILARITY " + str(sum(dist_list) / len(dist_list)))
        
    valence_means = []
    valence_stds = []
    arousal_means = []
    arousal_stds = []
    for word in valence_sem_groups[k]:
        word_row = bawl[bawl["WORD_LOWER"] == word]
        valence_means.append(float(word_row["EMO_MEAN"]))
        valence_stds.append(float(word_row["EMO_STD"]))
        arousal_means.append(float(word_row["AROUSAL_MEAN"]))
        arousal_stds.append(float(word_row["AROUSAL_STD"]))

    print(let + ": MEAN VALENCE " + str(sum(valence_means) / len(valence_means)))
    print(let + ": MEAN VALENCE STD " + str(sum(valence_stds) / len(valence_stds)))
    print(let + ": MEAN AROUSAL " + str(sum(arousal_means) / len(arousal_means)))
    print(let + ": MEAN AROUSAL STD " + str(sum(arousal_stds) / len(arousal_stds)))
    
    try:
        suggestion_value = (sum(dist_list) / len(dist_list) + abs(sum(valence_means) / len(valence_means)) - (sum(valence_stds) / len(valence_stds)))*(1.05**(len(np.array(high_valence_words)[class_members])))
    except:
        suggestion_value = 0
    suggestion_values.append((suggestion_value,np.array(high_valence_words)[class_members]))    
    print(let + ": SUGGESTION VALUE " + str(suggestion_values[-1][0]))

Ranking for suggestion values

suggestion_values = sorted(suggestion_values, key=lambda tup: tup[0], reverse=True)
for rank in range(len(suggestion_values)):
    print(str(rank+1) + " " + str(list(suggestion_values[rank][1])))

high_val_ffs = list(suggestion_values[0][1])
with open(path_pickled + "\\high_val_ffs.pkl", "wb") as f:
    pickle.dump(high_val_ffs, f)

#### 1.2.1b Low Valence

Define relevant valence words by threshold value

emo_mean = -2
emo_std = 1

low_valence_words = list(bawl[(bawl["FFS_FREQ"] >= 50) & (bawl["EMO_MEAN"] <= emo_mean) & (bawl["EMO_STD"] <= emo_std)]["WORD_LOWER"])

Define relevant valence words by ranking

bot_n = 25
emo_std = 1

low_valence_words = list(bawl[(bawl["FFS_FREQ"] >= 50) & (bawl["EMO_STD"] <= emo_std)].sort_values("EMO_MEAN")["WORD_LOWER"])[:bot_n]

Get vectors for valence words

valence_words_index = [model_ff.key_to_index[word] for word in low_valence_words]
valence_word_vectors = [model_ff.get_vector(i) for i in valence_words_index]

Normalize vectors on unit circle

valence_word_cosine = np.array(normalize(valence_word_vectors,norm='l2'))

Cluster with Affinity Propagation

valence_cos_cluster = AF(affinity='euclidean')
valence_cos_cluster.fit(valence_word_cosine)

Tranform cluster to dictionary

valence_sem_groups = {}
for i in range(len(valence_cos_cluster.labels_)):
    label = valence_cos_cluster.labels_[i]
    if label not in valence_sem_groups.keys():
        valence_sem_groups[label] = []

    word = low_valence_words[i]
    valence_sem_groups[label].append(word)
    
print(valence_sem_groups)

Print relevant information of clusters

suggestion_values = []

letters = cycle('abcdefghijklmnopqrstuvxyz')
for k, let in zip(range(len(valence_cos_cluster.cluster_centers_)), letters):
    class_members = valence_cos_cluster.labels_ == k

    dist_list = []
    print(let + ": ITEMS " + str(list(np.array(low_valence_words)[class_members])))
    for member1 in np.array(low_valence_words)[class_members]:
        for member2 in np.array(low_valence_words)[class_members]:
            if member1 != member2:
                dist_list.append(model_ff.similarity(member1,member2))
    if (dist_list != []):
        print(let + ": AVG SIMILARITY " + str(sum(dist_list) / len(dist_list)))
        
    valence_means = []
    valence_stds = []
    arousal_means = []
    arousal_stds = []
    for word in valence_sem_groups[k]:
        word_row = bawl[bawl["WORD_LOWER"] == word]
        valence_means.append(float(word_row["EMO_MEAN"]))
        valence_stds.append(float(word_row["EMO_STD"]))
        arousal_means.append(float(word_row["AROUSAL_MEAN"]))
        arousal_stds.append(float(word_row["AROUSAL_STD"]))

    print(let + ": MEAN VALENCE " + str(sum(valence_means) / len(valence_means)))
    print(let + ": MEAN VALENCE STD " + str(sum(valence_stds) / len(valence_stds)))
    print(let + ": MEAN AROUSAL " + str(sum(arousal_means) / len(arousal_means)))
    print(let + ": MEAN AROUSAL STD " + str(sum(arousal_stds) / len(arousal_stds)))
    
    try:
        suggestion_value = (sum(dist_list) / len(dist_list) + abs(sum(valence_means) / len(valence_means)) - (sum(valence_stds) / len(valence_stds)))*(1.05**(len(np.array(high_valence_words)[class_members])))
    except:
        suggestion_value = 0
    suggestion_values.append((suggestion_value,np.array(low_valence_words)[class_members]))    
    print(let + ": SUGGESTION VALUE " + str(suggestion_values[-1][0]))

Ranking for suggestion values

suggestion_values = sorted(suggestion_values, key=lambda tup: tup[0], reverse=True)
for rank in range(len(suggestion_values)):
    print(str(rank+1) + " " + str(list(suggestion_values[rank][1])))

low_val_ffs = list(suggestion_values[0][1])
with open(path_pickled + "\\low_val_ffs.pkl", "wb") as f:
    pickle.dump(low_val_ffs, f)

### 1.1.2 Arousal

#### 1.2.2a High Arousal

Define relevant arousal words by threshold value

arousal_mean = 2
arousal_std = 1

high_arousal_words = list(bawl[(bawl["FFS_FREQ"] >= 50) & (bawl["AROUSAL_MEAN"] >= arousal_mean) & (bawl["AROUSAL_STD"] <= arousal_std)]["WORD_LOWER"])

Define relevant arousal words by ranking

top_n = 25
arousal_std = 1

high_arousal_words = list(bawl[(bawl["FFS_FREQ"] >= 50) & (bawl["AROUSAL_STD"] <= arousal_std)].sort_values("AROUSAL_MEAN")["WORD_LOWER"])[-top_n:]

Get vectors for arousal words

arousal_words_index = [model_ff.key_to_index[word] for word in high_arousal_words]
arousal_word_vectors = [model_ff.get_vector(i) for i in arousal_words_index]

Normalize vectors on unit circle

arousal_word_cosine = np.array(normalize(arousal_word_vectors,norm='l2'))

Cluster with Affinity Propagation

arousal_cos_cluster = AF(affinity='euclidean')
arousal_cos_cluster.fit(arousal_word_cosine)

Tranform cluster to dictionary

arousal_sem_groups = {}
for i in range(len(arousal_cos_cluster.labels_)):
    label = arousal_cos_cluster.labels_[i]
    if label not in arousal_sem_groups.keys():
        arousal_sem_groups[label] = []

    word = high_arousal_words[i]
    arousal_sem_groups[label].append(word)
    
print(arousal_sem_groups)

Print relevant information of clusters

suggestion_values = []

letters = cycle('abcdefghijklmnopqrstuvxyz')
for k, let in zip(range(len(arousal_cos_cluster.cluster_centers_)), letters):
    class_members = arousal_cos_cluster.labels_ == k

    dist_list = []
    print(let + ": ITEMS " + str(list(np.array(high_arousal_words)[class_members])))
    for member1 in np.array(high_arousal_words)[class_members]:
        for member2 in np.array(high_arousal_words)[class_members]:
            if member1 != member2:
                dist_list.append(model_ff.similarity(member1,member2))
    if (dist_list != []):
        print(let + ": AVG SIMILARITY " + str(sum(dist_list) / len(dist_list)))
        
    valence_means = []
    valence_stds = []
    arousal_means = []
    arousal_stds = []
    for word in arousal_sem_groups[k]:
        word_row = bawl[bawl["WORD_LOWER"] == word]
        valence_means.append(float(word_row["EMO_MEAN"]))
        valence_stds.append(float(word_row["EMO_STD"]))
        arousal_means.append(float(word_row["AROUSAL_MEAN"]))
        arousal_stds.append(float(word_row["AROUSAL_STD"]))

    print(let + ": MEAN VALENCE " + str(sum(valence_means) / len(valence_means)))
    print(let + ": MEAN VALENCE STD " + str(sum(valence_stds) / len(valence_stds)))
    print(let + ": MEAN AROUSAL " + str(sum(arousal_means) / len(arousal_means)))
    print(let + ": MEAN AROUSAL STD " + str(sum(arousal_stds) / len(arousal_stds)))
    
    try:
        suggestion_value = (sum(dist_list) / len(dist_list) + sum(arousal_means) / len(arousal_means) - (sum(arousal_stds) / len(arousal_stds)))*(1.05**(len(np.array(high_arousal_words)[class_members])))
    except:
        suggestion_value = 0
    suggestion_values.append((suggestion_value,np.array(high_arousal_words)[class_members]))    
    print(let + ": SUGGESTION VALUE " + str(suggestion_values[-1][0]))

Ranking for suggestion values

suggestion_values = sorted(suggestion_values, key=lambda tup: tup[0], reverse=True)
for rank in range(len(suggestion_values)):
    print(str(rank+1) + " " + str(list(suggestion_values[rank][1])))

high_arousal_ffs = list(suggestion_values[0][1])
with open(path_pickled + "\\high_arousal_ffs.pkl", "wb") as f:
    pickle.dump(high_arousal_ffs, f)

#### 1.2.2b Low Arousal

Define relevant arousal words by threshold value

arousal_mean = 2
arousal_std = 1

low_arousal_words = list(bawl[(bawl["FFS_FREQ"] >= 50) & (bawl["AROUSAL_MEAN"] <= arousal_mean) & (bawl["AROUSAL_STD"] <= arousal_std)]["WORD_LOWER"])

Define relevant arousal words by ranking

bot_n = 25
arousal_std = 1

low_arousal_words = list(bawl[(bawl["FFS_FREQ"] >= 50) & (bawl["AROUSAL_STD"] <= arousal_std)].sort_values("AROUSAL_MEAN")["WORD_LOWER"])[:bot_n]

Get vectors for arousal words

arousal_words_index = [model_ff.key_to_index[word] for word in low_arousal_words]
arousal_word_vectors = [model_ff.get_vector(i) for i in arousal_words_index]

Normalize vectors on unit circle

arousal_word_cosine = np.array(normalize(arousal_word_vectors,norm='l2'))

Cluster with Affinity Propagation

arousal_cos_cluster = AF(affinity='euclidean')
arousal_cos_cluster.fit(arousal_word_cosine)

Tranform cluster to dictionary

arousal_sem_groups = {}
for i in range(len(arousal_cos_cluster.labels_)):
    label = arousal_cos_cluster.labels_[i]
    if label not in arousal_sem_groups.keys():
        arousal_sem_groups[label] = []

    word = low_arousal_words[i]
    arousal_sem_groups[label].append(word)
    
print(arousal_sem_groups)

Print relevant information of clusters

suggestion_values = []

letters = cycle('abcdefghijklmnopqrstuvxyz')
for k, let in zip(range(len(arousal_cos_cluster.cluster_centers_)), letters):
    class_members = arousal_cos_cluster.labels_ == k

    dist_list = []
    print(let + ": ITEMS " + str(list(np.array(low_arousal_words)[class_members])))
    for member1 in np.array(low_arousal_words)[class_members]:
        for member2 in np.array(low_arousal_words)[class_members]:
            if member1 != member2:
                dist_list.append(model_ff.similarity(member1,member2))
    if (dist_list != []):
        print(let + ": AVG SIMILARITY " + str(sum(dist_list) / len(dist_list)))
        
    valence_means = []
    valence_stds = []
    arousal_means = []
    arousal_stds = []
    for word in arousal_sem_groups[k]:
        word_row = bawl[bawl["WORD_LOWER"] == word]
        valence_means.append(float(word_row["EMO_MEAN"]))
        valence_stds.append(float(word_row["EMO_STD"]))
        arousal_means.append(float(word_row["AROUSAL_MEAN"]))
        arousal_stds.append(float(word_row["AROUSAL_STD"]))

    print(let + ": MEAN VALENCE " + str(sum(valence_means) / len(valence_means)))
    print(let + ": MEAN VALENCE STD " + str(sum(valence_stds) / len(valence_stds)))
    print(let + ": MEAN AROUSAL " + str(sum(arousal_means) / len(arousal_means)))
    print(let + ": MEAN AROUSAL STD " + str(sum(arousal_stds) / len(arousal_stds)))
    
    try:
        suggestion_value = (sum(dist_list) / len(dist_list) + (4 - sum(arousal_means) / len(arousal_means)) - (sum(arousal_stds) / len(arousal_stds)))*(1.05**(len(np.array(low_arousal_words)[class_members])))
    except:
        suggestion_value = 0
    suggestion_values.append((suggestion_value,np.array(low_arousal_words)[class_members]))    
    print(let + ": SUGGESTION VALUE " + str(suggestion_values[-1][0]))

Ranking for suggestion values

suggestion_values = sorted(suggestion_values, key=lambda tup: tup[0], reverse=True)
for rank in range(len(suggestion_values)):
    print(str(rank+1) + " " + str(list(suggestion_values[rank][1])))

low_arousal_ffs = list(suggestion_values[0][1])
with open(path_pickled + "\\low_arousal_ffs.pkl", "wb") as f:
    pickle.dump(low_arousal_ffs, f)
